Well, we have completed our journey!!农业机械化无敌!
The only thing left to do now is practice,practice, practice.Even though we covered a lot of ground in our trek, we barely scratched the surface as far as the command line goes.There are still thousands of command line programs left to be discovered and enjoyed. Start diggin around in /usr/bin and you'll see!